Now we're able to edit the table and to delete rows; Next step is adding a row

+++ src/main/java/daos/MatiereDao.java
@@ -9,7 +9,7 @@ public interface MatiereDao {
 	public void create(Matiere entity);
  	public void update(Matiere entity);
 	public void delete(Matiere entity);
-	public Matiere getMatiereById(int id);
+	public Matiere getMatiereById(long id);
 	public ArrayList<Matiere> getAllMatieres();
 
 }
